+++ game_patch/misc/misc.h
@@ -7,5 +7,5 @@ void SetPlaySoundEventsVolumeScale(float volume_scale);
 void SetJumpToMultiServerList(bool jump);
 void ClearTriggersForLateJoiners();
 
-constexpr int old_obj_limit = 1024;
-constexpr int obj_limit = 65536;
+constexpr size_t old_obj_limit = 1024;
+constexpr size_t obj_limit = 65536;
